requirements.txt
altair==4.2.2
anyio==3.6.2
appnope==0.1.3
asn1crypto==1.5.1
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
bcrypt==4.0.1
blinker==1.5
boto3==1.26.69
botocore==1.29.69
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
comm==0.1.2
contourpy==1.0.7
cryptography==39.0.1
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
dnspython==2.3.0
email-validator==1.3.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
extra-streamlit-components==0.1.56
fastapi==0.92.0
filelock==3.9.0
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.30
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.21.2
ipython==8.10.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
jupyter_client==8.0.2
jupyter_core==5.2.0
jwt==1.3.1
kiwisolver==1.4.4
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.5.6
numpy==1.24.2
openpyxl==3.1.0
orjson==3.8.6
oscrypto==1.3.0
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
plotly==5.13.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pycparser==2.21
pycryptodomex==3.17
pydantic==1.10.5
pydeck==0.8.0
Pygments==2.14.0
PyJWT==2.6.0
Pympler==1.0.1
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.7
python-dotenv==0.21.1
python-multipart==0.0.5
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.0.0
requests==2.28.2
rfc3986==1.5.0
rich==13.3.1
s3transfer==0.6.0
seaborn==0.12.2
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowflake-connector-python==3.0.0
stack-data==0.6.2
starlette==0.25.0
streamlit==1.18.1
streamlit-authenticator==0.1.5
streamlit-option-menu==0.3.2
tenacity==8.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.9.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
ujson==5.7.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
validators==0.20.0
watchfiles==0.18.1
wcwidth==0.2.6
zipp==3.13.0
websockets==10.4